Move branding under environments/ directory.
